requirements.txt
streamlit
onnxruntime
requests
numpy
pandas
opencv-python-headless
scikit-learn
Pillow
matplotlib
tensorflow
torch